fix(web): allow direct backend dev origin
