+++ web/next.config.ts
@@ -5,6 +5,7 @@ import { withSentryConfig } from "@sentry/nextjs";
 import path from "node:path";
 import { fileURLToPath } from "node:url";
 import { poweredByHeader, securityHeaderRules } from "./security-headers";
+import { directDevBackendHost } from "./app/lib/direct-dev-backend-origin";
 
 const withNextIntl = createNextIntlPlugin("./i18n/request.ts");
 const webRoot = path.dirname(fileURLToPath(import.meta.url));
@@ -14,6 +15,11 @@ const releaseDocsOrigin =
   process.env.CMUX_RELEASE_DOCS_ORIGIN ?? "https://cmux-docs-release.vercel.app";
 const nightlyDocsOrigin =
   process.env.CMUX_NIGHTLY_DOCS_ORIGIN ?? "https://cmux-docs-nightly.vercel.app";
+// The embedded browser reaches a dev server through its per-instance
+// Tailscale Serve hostname. Next.js blocks cross-origin HMR and RSC resources
+// unless that hostname is explicitly allowed. Keep this opt-in and validated
+// so production and SSH-backed development retain the default protection.
+const directDevBackendAllowedHost = directDevBackendHost();
 
 // Agent landing pages moved under /agents/<agent>. Keep the old top-level
 // slugs working with permanent redirects, for the bare English path and every
@@ -56,6 +62,9 @@ const tuiInstallerHeaderRules = [
 
 const nextConfig: NextConfig = {
   poweredByHeader,
+  allowedDevOrigins: directDevBackendAllowedHost
+    ? [directDevBackendAllowedHost]
+    : undefined,
   cacheComponents: true,
   partialPrefetching: true,
   experimental: {
